requirements.txt
mock==1.0.1
requests==2.2.1
responses==0.3.0
six==1.9.0
python-dateutil==1.5